some things I forgot to commit

+++ src/types/ctr.nim
@@ -30,15 +30,21 @@ proc getConsoleName*(self: Ctr): string = "Nintendo 3DS"
 proc getElfBinaryName*(self: Ctr): string = "3DS.elf"
 proc getIconExtension*(self: Ctr): string = "png"
 
-proc shouldConvertFile(self: Ctr, source: string, destination: string): bool =
-    if not fileExists(destination) or fileNewer(source, destination):
-        logger.info(fmt"Converting file {source} to {destination}!")
-        return true
+proc shouldHandleFile(self: Ctr, source: string, destination: string, ext: string): bool =
+    let checked_file = fmt"{destination}.{ext}"
 
+    try:
+        if not fileExists(checked_file) or fileNewer(source, checked_file):
+            logger.info(fmt"Copying/Converting file {source} to {checked_file}.")
+            return true
+    except Exception as e:
+        logger.error(fmt"Something went wrong: {e.msg} ({checked_file})")
+
+    logger.info(fmt"File source {source} for {checked_file} is not newer. Skipping.")
     return false
 
 proc convertFiles(self: Ctr, source: string): bool =
-    echo(strings.ConvertCopyingFiles)
+    stdout.write(strings.ConvertCopyingFiles)
 
     let romFS = console.getRomFSDirectory()
 
@@ -48,16 +54,20 @@ proc convertFiles(self: Ctr, source: string): bool =
 
         let (dir, name, extension) = os.splitFile(path)
 
-        let relativePath = fmt("{source}/{path}")
+        let relativePath = normalizedPath(fmt("{source}/{path}"))
         let destination = fmt("{romFS}/{dir}")
+
+        var destinationPath: string = ""
 
         try:
             os.createDir(destination)
 
-            let destinationPath = fmt("{destination}/{name}")
+            destinationPath = normalizedPath(fmt("{destination}/{name}"))
 
             if extension in Textures or extension in Fonts:
-                if not self.shouldConvertFile(relativePath, destinationPath):
+                let ext = if extension in Textures: "t3x" else: "bcfnt"
+
+                if not self.shouldHandleFile(relativePath, destinationPath, ext):
                     continue
 
                 var conversion_command: string = ""
@@ -66,27 +76,33 @@ proc convertFiles(self: Ctr, source: string): bool =
                     conversion_command = TextureCommand.format(relativePath, destinationPath)
                 elif extension in Fonts:
                     conversion_command = FontCommand.format(relativePath, destinationPath)
-                else:
-                    os.copyFileToDir(relativePath, destination)
 
                 if not conversion_command.isEmptyOrWhitespace():
-                    runCommand(conversion_command)
+                    if not runCommand(conversion_command):
+                        return false
+            else:
+                if not self.shouldHandleFile(relativePath, destinationPath, extension.substr(1)):
+                    continue
+
+                os.copyFileToDir(relativePath, destination)
         except Exception as e:
+            logger.error(fmt"Copy or conversion error! {e.msg}: {relativePath} -> {destination}/{destinationPath}")
             return false
 
+    echo("Done!")
     return true
 
 proc publish*(self: Ctr, source: string): bool =
+    logger.info(fmt"== [{self.getConsoleName()}] ==")
+
     if not self.convertFiles(source):
-        return
+        return false
 
     let elfBinaryPath = self.getElfBinaryPath()
 
     if not os.fileExists(elfBinaryPath) and not config.rawData:
-        echo(strings.ElfBinaryNotFound.format(
-                config.name, self.getConsoleName(), self.getElfBinaryName(),
-                config.binSearchPath))
-        return
+        echo(strings.ElfBinaryNotFound.format(config.name, self.getConsoleName(), self.getElfBinaryName(), config.binSearchPath))
+        return false
 
     let properDescription = fmt("{config.description} • {config.version}")
     let outputPath = self.getGenericOutputBinaryPath()
@@ -96,17 +112,20 @@ proc publish*(self: Ctr, source: string): bool =
 
         # Copy RomFS graphics content to directory
         for name, content in CtrGraphics.items():
-            writeFile(fmt"{RomFSDirectory}/{name}", content)
+            if not fileExists(fmt"{RomFSDirectory}/{name}"):
+                writeFile(fmt"{RomFSDirectory}/{name}", content)
+            else:
+                logger.info(fmt"Messagebox texture '{name}' already exists. Skipping.")
 
         let (head, _) = splitPath(RomFSDirectory)
 
         # Output {SuperGame}.smdh to `build` directory
-        console.runCommand(SmdhCommand.format(config.name, properDescription,
-                config.author, self.getIcon(), outputPath))
+        if not console.runCommand(SmdhCommand.format(config.name, properDescription, config.author, self.getIcon(), outputPath)):
+            return false
 
         # Output {SuperGame}.3dsx to `build` directory
-        console.runCommand(BinaryCommand.format(self.getElfBinaryPath(),
-                outputPath, head))
+        if not console.runCommand(BinaryCommand.format(self.getElfBinaryPath(), outputPath, head)):
+            return false
     except Exception as e:
         logger.error(fmt"{self.getConsoleName()} publishing failure: {e.msg}")
         return false
